# feat: implement SPEC-11 API performance optimizations