--- a/src/basic_memory/mcp/tools/__init__.py
+++ b/src/basic_memory/mcp/tools/__init__.py
@@ -24,6 +24,7 @@ from basic_memory.mcp.tools.project_management import (
     create_memory_project,
     delete_project,
 )
+
 # ChatGPT-compatible tools
 from basic_memory.mcp.tools.chatgpt_tools import search, fetch